fix properties

+++ src/test/java/fr/eni/GestionPotager/bo/CarreTest.java
@@ -65,7 +65,7 @@ class CarreTest {
 	void testSetSurface() {
 		Potager potager = new Potager("STR", "STR", 25, "STR");
 		Carre carre1 = new Carre(2, "sableux", Exposition.MI_OMBRE, potager);
-		carre1.setSurface(12);
+		carre1.setSurface((double) 12);
 		assertEquals(12, carre1.getSurface());
 	}
